Problem003

+++ ProjectEuler.ConsoleApp/Program.cs
@@ -7,7 +7,7 @@ namespace ProjectEuler.ConsoleApp
         static void Main(string[] args)
         {
             Console.Write("Input number : ");
-            var number = int.Parse(Console.ReadLine());
+            var number = double.Parse(Console.ReadLine());
             var calculate = new Calculate();
             var result = calculate.Calc(number);
             Console.Write($"Result = {result}");
